// src/city/Car/interfaces.ts
import {Coords, Size} from '../interfaces';

export interface CarOptions {
    size: Size;
    coords: Coords;
    animation: {
        endCoords: Coords;
        speedTime: number;
    };
    materials: {
        color: string;
    };
    name: string;
}

export interface CarData {
    name: string;
    color: string;
    coords: Coords;
    animation: {
        endCoords: {x?: number; y?: number; z?: number};
        speedTime: number;
    };
}